requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
audioread==3.0.1
billiard==4.2.1
blis==1.3.0
catalogue==2.0.10
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.21.1
confection==0.1.5
cryptography==45.0.3
cymem==2.0.11
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.115.12
grpcio==1.73.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.5.4
langchain==0.3.25
langchain-core==0.3.64
langchain-text-splitters==0.3.8
langcodes==3.5.0
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.45
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.43.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
murmurhash==1.0.13
numba==0.60.0
numpy==2.0.2
openai==1.85.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
portalocker==2.10.1
preshed==3.0.10
prompt_toolkit==3.0.51
protobuf==6.31.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
qdrant-client==1.14.2
redis==6.2.0
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
scikit-learn==1.6.1
scipy==1.13.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
starlette==0.46.2
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wrapt==1.17.2
xxhash==3.5.0
zstandard==0.23.0
